requirements.txt
langchain==1.0.1
langchain-core==1.0.1
langchain-community
langchain-text-splitters
langchain-google-genai
gradio
PyMuPDF
python-dotenv
faiss-cpu
